pyproject.toml
[project]
name = "spacex-landing"
version = "0.1.0"
description = "SpaceX-style rocket landing simulator using MuJoCo and SAC reinforcement learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torchrl>=0.6.0",
    "tensordict>=0.6.0",
    "torchvision>=0.19.0",
    "mujoco>=3.2.0",
    "gymnasium>=1.0.0",
    "imageio>=2.36.0",
    "imageio-ffmpeg>=0.5.1",
    "hydra-core>=1.3.0",
    "wandb>=0.18.0",
    "tqdm>=4.67.0",
    "moviepy>=2.1.0",
]

[project.optional-dependencies]
gpu = [
    "mujoco-warp",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["env"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
